+++ smallworld/emulators/angr.py
@@ -174,7 +174,7 @@ class AngrEmulator(emulator.Emulator):
             raise PathTerminationSignal()
 
         # Set breakpoints to halt on exit
-        exits = list(code.exits)
+        exits = [b.stop for b in code.bounds]
         default_exit = code.base + len(code.image)
         if code.type == "blob" and default_exit not in exits:
             # Set a default exit point to keep us from
@@ -225,7 +225,8 @@ class AngrEmulator(emulator.Emulator):
         self.mgr.move(
             from_stash="active",
             to_stash="deadended",
-            filter_func=lambda x: x._ip.concrete_value in self._code.exits,
+            filter_func=lambda x: x._ip.concrete_value
+            in [b.stop for b in self._code.bounds],
         )
 
         # Test for exceptional states